fixed string allocation to allocate only the size of string_buf. Added function declaration for print token function. Changed yylineno type to int.

+++ tokens.hpp
@@ -49,6 +49,6 @@ extern char* STR_BUF;
 
   extern int yylineno;
   extern char* yytext;
-  extern unsigned long yyleng; // instead of int
+  extern int yyleng;
   extern int yylex();
 #endif /* TOKENS_HPP_ */
